Rename TUMLiveVjs to player
